requirements.txt
psycopg2-binary==2.9.1
Flask==2.0.2
requests==2.25.1
google-auth==2.3.3
Flask-Login==0.5.0